--- product-scenarios/1-user-self-registration-to-an-application/1.1-user-registration-with-web-application-itself/README.md
@@ -1,9 +1,0 @@
-# 1.1 User registration provided in the web application itself
-
-Using provisioning APIs in identity server, trigger user creation in IS from the web application. 
-
-# Test Scenarios
-You can provision user into identity server in the following ways. 
-
-- [1.1.1 Provision user in identity server using SCIM2 API](1.1.1-provision-user-using-SCIM2/README.md)
-- [1.1.2 Provision user in identity server using SCIM1.1 API](1.1.2-provision-user-using-SCIM1.1/README.md)
